Can't seem to make up my mind, can I?
- Renamed assembly to Foundry.Disassembly.Mips for generic use
- Vastly simplified the process while I am in the middle of learning. Tree building and converting to other languages will be implemented once the MIPS disassembling is perfect.
- Fixed a glaring issue where the runtime was converting big ending uints to little endian behind the scenes during its reinterpret cast. This was giving totally incorrect byte values and therefore invalid instructions. BitConverter.ToUInt32 usage was replaced with BinaryPrimitives.ReadUint32BigEndian to force big endian values.
- Introduced a bunch of regressions.

The next action items are to fill in the remaining operations, reintroduce switch case label detection, and reintroduce the 2-pass system for smarter disassembly of methods, variables, and labels.

In addition, I need to cross reference the code I have with multiple other MIPS disassemblers to fill in any gaps and verify that the logic I have is correct.